added code for error calculation on the Lacy/Donley plots
Added reset of LaTeX code for the error propagation, this is now fully describes the error in the x and y axis for the lacy/donley plots. The error is now also calculated in a function and added to the overall dataset, the error corresponds to two columns which ae the error in each axis.